apps/first_challengue/requirements.txt
# Function dependencies, for example:
google-cloud-bigquery
google-cloud-bigquery-storage
flask
